@@ -1,23 +1,21 @@
 {
-  "name": "omatech/mcore",
+  "name": "omatech/mage-core",
   "require": {
-    "php": "^8.0",
+    "php": "^8.2",
     "lambdish/phunctional": "v2.1",
     "webonyx/graphql-php": "^v14.11"
   },
   "require-dev": {
+    "infection/infection": "^0.26",
+    "mockery/mockery": "^1.5",
+    "nunomaduro/phpinsights": "^v2.6",
+    "phpunit/phpunit": "^9.5",
     "roave/security-advisories": "dev-latest",
-    "phpunit/phpunit": "^9.0",
-    "infection/infection": "^0",
-    "nunomaduro/phpinsights": "^2.0",
-    "symfony/var-dumper": "^6.0",
-    "mockery/mockery": "^1.4",
-    "fakerphp/faker": "^1.0",
-    "phpstan/phpstan": "^1.3"
+    "symfony/var-dumper": "^v6.1"
   },
   "autoload": {
     "psr-4": {
-      "Omatech\\Mcore\\": "src/"
+      "Omatech\\MageCore\\": "src/"
     }
   },
   "autoload-dev": {
@@ -32,11 +30,7 @@
       "@phpunit",
       "@infection"
     ],
-    "phpinsights": "vendor/bin/phpinsights analyse --verbose --config-path=rules.php --no-interaction --fix src",
-    "phpstan": "vendor/bin/phpstan analyse -l 6 src",
-    "analyse": [
-      "@phpinsights"
-    ]
+    "analyse": "vendor/bin/phpinsights analyse --verbose --config-path=rules.php --no-interaction --fix src"
   },
   "config": {
     "optimize-autoloader": true,
