@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Value;
+namespace Omatech\MageCore\Editora\Domain\Value;
 
 use function Lambdish\Phunctional\each;
 use function Lambdish\Phunctional\map;
@@ -9,7 +9,7 @@ use function Lambdish\Phunctional\search;
 final class ValueCollection
 {
     /** @var array<BaseValue> */
-    private array $values;
+    private readonly array $values;
 
     /** @param array<BaseValue> $values */
     public function __construct(array $values)
@@ -20,17 +20,13 @@ final class ValueCollection
     public function fill(array $values): void
     {
         each(function (mixed $value): void {
-            search(static function (BaseValue $fillableValue) use ($value): bool {
-                return $fillableValue->language() === $value['language'];
-            }, $this->values)?->fill($value);
+            search(static fn (BaseValue $fillableValue): bool => $fillableValue->language() === $value['language'], $this->values)?->fill($value);
         }, $values);
     }
 
     public function language(string $language): ?BaseValue
     {
-        return search(static function (BaseValue $baseValue) use ($language) {
-            return $baseValue->language() === $language;
-        }, $this->values);
+        return search(static fn (BaseValue $baseValue) => $baseValue->language() === $language, $this->values);
     }
 
     /** @return array<BaseValue> */
