@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Attribute\Contracts;
+namespace Omatech\MageCore\Editora\Domain\Attribute\Contracts;
 
 interface AttributeRepositoryInterface
 {
