@@ -1,12 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Contracts;
+namespace Omatech\MageCore\Editora\Domain\Extraction\Contracts;
 
-use Omatech\Mcore\Editora\Domain\Extraction\Results;
+use Omatech\MageCore\Editora\Domain\Extraction\Results;
 
-interface ExtractionRepositoryInterface
+interface ExtractionInterface
 {
     public function instancesBy(array $params): Results;
-
     public function findRelatedInstances(string $instanceUuid, array $params): Results;
 }
