@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Extraction;
+namespace Omatech\MageCore\Editora\Domain\Extraction;
 
 use DateTime;
 use function Lambdish\Phunctional\first;
@@ -9,10 +9,9 @@ use function Lambdish\Phunctional\reduce;
 
 final class Extraction
 {
-    private string $query;
-    private string $hash;
-    private DateTime $date;
-    /** @var array<Query> $queries */
+    private readonly string $query;
+    private readonly string $hash;
+    private readonly DateTime $date;
     private array $queries;
 
     public function __construct(string $query)
@@ -37,13 +36,11 @@ final class Extraction
         return $this->date;
     }
 
-    /** @return array<Query> $queries */
     public function queries(): array
     {
         return $this->queries;
     }
 
-    /** @param array<Query> $queries */
     public function setQueries(array $queries): void
     {
         $this->queries = $queries;
@@ -52,12 +49,13 @@ final class Extraction
     public function toArray(): array
     {
         $results = reduce(static function (array $acc, Query $query): array {
-            $instances = map(static function (Instance $instance): array {
-                return $instance->toArray();
-            }, $query->results());
+            $instances = map(
+                static fn (Instance $instance): array => $instance->toArray(),
+                $query->results()
+            );
             $acc[] = count($instances) < 2 ? first($instances) ?? [] : $instances;
             return $acc;
         }, $this->queries, []);
-        return count($results) < 2 ? first($results) ?? [] : $results;
+        return (is_countable($results) ? count($results) : 0) < 2 ? first($results) ?? [] : $results;
     }
 }
