@@ -1,25 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Extraction;
+namespace Omatech\MageCore\Editora\Domain\Extraction;
 
 use function Lambdish\Phunctional\map;
 use function Lambdish\Phunctional\reduce;
 
 final class Query
 {
-    /** @var array|mixed */
-    private array $params;
-
-    /** @var array<Attribute> */
-    private array $attributes;
-
-    /** @var array<Query> $relations */
-    private array $relations;
-
-    /** @var Pagination|null */
+    private readonly array $params;
+    private readonly array $attributes;
+    private readonly array $relations;
     private ?Pagination $pagination = null;
-
-    /** @var array<Instance> $results */
     private array $results;
 
     public function __construct(array $query)
@@ -29,13 +20,11 @@ final class Query
         $this->relations = $query['relations'];
     }
 
-    /** @return array<Attribute> */
     public function attributes(): array
     {
         return $this->attributes;
     }
 
-    /** @return array<Query> */
     public function relations(): array
     {
         return $this->relations;
@@ -46,13 +35,11 @@ final class Query
         return $this->params;
     }
 
-    /** @return array<Instance> */
     public function results(): array
     {
         return $this->results;
     }
 
-    /** @param array<Instance> $results */
     public function setResults(array $results): self
     {
         $this->results = $results;
@@ -74,9 +61,7 @@ final class Query
     {
         return [
             'languages' => $this->languages(),
-            'attributes' => map(static function (Attribute $attribute): array {
-                return $attribute->toQuery();
-            }, $this->attributes),
+            'attributes' => map(static fn (Attribute $attribute): array => $attribute->toQuery(), $this->attributes),
             'params' => $this->params,
             'relations' => reduce(static function (array $acc, Query $query): array {
                 $acc[] = $query->toArray();
