@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Exceptions;
+namespace Omatech\MageCore\Editora\Domain\Instance\Exceptions;
 
 use Exception;
 
