@@ -1,11 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance;
+namespace Omatech\MageCore\Editora\Domain\Instance;
 
 final class InstanceRelation
 {
-    private string $key;
-    private array $instances;
+    private readonly string $key;
+    private readonly array $instances;
 
     public function __construct(string $key, array $instances)
     {
