@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance;
+namespace Omatech\MageCore\Editora\Domain\Instance;
 
 use DateTime;
-use Omatech\Mcore\Editora\Domain\Instance\Exceptions\InvalidEndDatePublishingException;
+use Omatech\MageCore\Editora\Domain\Instance\Exceptions\InvalidEndDatePublishingException;
 
 final class Publication
 {
