@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Validator\Rules;
+namespace Omatech\MageCore\Editora\Domain\Instance\Validator\Rules;
 
-use Omatech\Mcore\Editora\Domain\Attribute\AttributeCollection;
-use Omatech\Mcore\Editora\Domain\Value\BaseValue;
+use Omatech\MageCore\Editora\Domain\Attribute\AttributeCollection;
+use Omatech\MageCore\Editora\Domain\Value\BaseValue;
 
 abstract class BaseRule
 {
