@@ -1,16 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Validator\Rules;
+namespace Omatech\MageCore\Editora\Domain\Instance\Validator\Rules;
 
-use Omatech\Mcore\Editora\Domain\Instance\Validator\Exceptions\RequiredValueException;
-use Omatech\Mcore\Editora\Domain\Value\BaseValue;
-use Omatech\Mcore\Shared\Utils\Utils;
+use Omatech\MageCore\Editora\Domain\Instance\Validator\Exceptions\RequiredValueException;
+use Omatech\MageCore\Editora\Domain\Value\BaseValue;
+use Omatech\MageCore\Shared\Utils\Utils;
 
 final class Required extends BaseRule
 {
     public function validate(BaseValue $value): void
     {
-        if ($this->conditions === true && Utils::getInstance()->isEmpty($value->value())) {
+        if ($this->conditions === true && Utils::isEmpty($value->value())) {
             RequiredValueException::withValue($value);
         }
     }
