@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Validator\Rules;
+namespace Omatech\MageCore\Editora\Domain\Instance\Validator\Rules;
 
-use Omatech\Mcore\Editora\Domain\Attribute\Attribute;
-use Omatech\Mcore\Editora\Domain\Instance\Validator\Exceptions\UniqueValueException;
-use Omatech\Mcore\Editora\Domain\Value\BaseValue;
+use Omatech\MageCore\Editora\Domain\Attribute\Attribute;
+use Omatech\MageCore\Editora\Domain\Instance\Validator\Exceptions\UniqueValueException;
+use Omatech\MageCore\Editora\Domain\Value\BaseValue;
 use function Lambdish\Phunctional\filter;
 use function Lambdish\Phunctional\flat_map;
 
@@ -17,11 +17,9 @@ class Unique extends BaseRule
 
     private function validateInInstance(BaseValue $value): void
     {
-        $results = filter(static function (BaseValue $current) use ($value): bool {
-            return $current->value() === $value->value() &&
-                $current->value() !== null &&
-                $value->value() !== null;
-        }, $this->attributesValues($value->key()));
+        $results = filter(static fn (BaseValue $current): bool => $current->value() === $value->value() &&
+            $current->value() !== null &&
+            $value->value() !== null, $this->attributesValues($value->key()));
 
         if (count($results) > 1) {
             UniqueValueException::withValue($value);
@@ -30,8 +28,6 @@ class Unique extends BaseRule
 
     private function attributesValues(string $key): array
     {
-        return flat_map(static function (Attribute $attribute): array {
-            return $attribute->values()->get();
-        }, $this->attributeCollection->findAll($key));
+        return flat_map(static fn (Attribute $attribute): array => $attribute->values()->get(), $this->attributeCollection->findAll($key));
     }
 }
