@@ -1,15 +1,15 @@
 <?php declare(strict_types=1);
 
-namespace Omatech\Mcore\Editora\Domain\Instance\Validator\Rules;
+namespace Omatech\MageCore\Editora\Domain\Instance\Validator\Rules;
 
-use Omatech\Mcore\Editora\Domain\Attribute\AttributeCollection;
-use Omatech\Mcore\Editora\Domain\Instance\Validator\Contracts\UniqueValueInterface;
-use Omatech\Mcore\Editora\Domain\Instance\Validator\Exceptions\UniqueValueInDBException;
-use Omatech\Mcore\Editora\Domain\Value\BaseValue;
+use Omatech\MageCore\Editora\Domain\Attribute\AttributeCollection;
+use Omatech\MageCore\Editora\Domain\Instance\Validator\Contracts\UniqueValueInterface;
+use Omatech\MageCore\Editora\Domain\Instance\Validator\Exceptions\UniqueValueInDBException;
+use Omatech\MageCore\Editora\Domain\Value\BaseValue;
 
 final class UniqueDB extends Unique
 {
-    private UniqueValueInterface $repository;
+    private readonly UniqueValueInterface $repository;
 
     public function __construct(AttributeCollection $attributeCollection, mixed $conditions)
     {
