@@ -1,48 +1,34 @@
-<?php declare(strict_types=1);
+<?php
 
 namespace Tests\Editora\Data\Objects;
 
-use function Lambdish\Phunctional\reduce;
+use Tests\Editora\Data\InstanceArrayBuilder;
+use Tests\Editora\Data\InstanceFactory;
 
-class PhotosMother extends ObjectMother
+class PhotosMother
 {
-    protected array $availableRelations = [
-        'photos-locations' => [
-            LocationsMother::class,
-            CoordinatesMother::class,
-        ],
-    ];
-
-    public function get(int $instancesNumber = 1, ?string $key = null, ?array $relations = []): array
+    public static function get(int $instancesNumber, array $relations = []): array
     {
-        $this->instances = [];
-        for ($i = 1; $i <= $instancesNumber; ++$i) {
-            $this->instances[] = $this->build('Photos')->fill([
-                'metadata' => [
-                    'uuid' => $this->faker->uuid(),
-                    'key' => $key ?? 'photo-instance-' . $i,
-                    'publication' => [
-                        'startPublishingDate' => $this->faker->dateTime()->format('Y-m-d H:i:s'),
-                    ],
-                ],
-                'attributes' => [
-                    'url' => [
-                        'values' => reduce(function (array $acc, string $language) {
-                            $acc[] = [
-                                'uuid' => $this->faker->uuid(),
-                                'language' => $language,
-                                'value' => $this->faker->url(),
-                            ];
-
-                            return $acc;
-                        }, $this->languages, []),
-                        'attributes' => [],
-                    ],
-                ],
-                'relations' => $relations,
-            ]);
+        $instances = [];
+        for($i=1; $i <= $instancesNumber; $i++) {
+            $instances[] = InstanceFactory::fill('Photos', static function (InstanceArrayBuilder $builder) use ($relations, $i) {
+                $builder
+                    ->addClassRelations('photos-locations', [ 'locations', 'coordinates' ])
+                    ->addMetadata('uuid', 'photo-instance-'.$i)
+                    ->addAttribute('url', 'string', [
+                        ['uuid' => 'uuid', 'language' => 'es', 'value' => 'url-es-'.$i],
+                        ['uuid' => 'uuid', 'language' => 'en', 'value' => 'url-en-'.$i]
+                    ]);
+                foreach($relations as $relation => $relatedInstances) {
+                    $relatedInstancesIds = [];
+                    foreach($relatedInstances as $relatedInstance) {
+                        $relatedInstancesIds[$relatedInstance->uuid()] = $relatedInstance->data()['classKey'];
+                    }
+                    $builder->addRelation($relation, $relatedInstancesIds);
+                }
+                return $builder->build();
+            });
         }
-
-        return $this->instances;
+        return $instances;
     }
 }
